Split ops_tags.rs into operations.rs and tags.rs (aterm-8tn.37) (#194)
Pure refactor: move Operation + list_operations into src/ashirt/operations.rs
and Tag, NewTag, list_tags, create_tag, random_tag_color, TAG_COLORS into
src/ashirt/tags.rs, each with its own tests. Update mod.rs to declare the two
modules instead of ops_tags, and repoint call sites in menu.rs and
upload_menu.rs. No behavior change.
