+++ src/ashirt/tags.rs
@@ -1,8 +1,7 @@
-//! ASHIRT operations and tags API.
+//! ASHIRT tags API.
 //!
-//! Port of Go `network/get_operations.go` + `network/tagging.go`. Builds on the
-//! signed JSON helpers from [`crate::ashirt::http`]:
-//!   * [`list_operations`]  — `GET /operations`
+//! Port of Go `network/tagging.go`. Builds on the signed JSON helpers from
+//! [`crate::ashirt::http`]:
 //!   * [`list_tags`]        — `GET /operations/{slug}/tags`
 //!   * [`create_tag`]       — `POST /operations/{slug}/tags`
 //!   * [`random_tag_color`] — pick a color from the ASHIRT tag palette
@@ -16,28 +15,6 @@ use std::hash::{BuildHasher, Hasher};
 use serde::{Deserialize, Serialize};
 
 use super::http::{Client, HttpError};
-
-/// An ASHIRT operation, as returned by `GET /operations`.
-///
-/// Only the fields aterm consumes are modeled; serde ignores any additional
-/// fields the server may include, so the struct stays forward-compatible.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Operation {
-    /// URL-safe identifier used to address the operation in API paths.
-    pub slug: String,
-    /// Human-readable operation name.
-    pub name: String,
-    /// Numeric operation id. Defaulted when absent so partial payloads still
-    /// deserialize.
-    #[serde(default)]
-    pub id: i64,
-    /// Operation status code. Defaulted when absent.
-    #[serde(default)]
-    pub status: i64,
-    /// Number of users with access to the operation. Defaulted when absent.
-    #[serde(default, rename = "numUsers")]
-    pub num_users: i64,
-}
 
 /// A tag within an operation.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -82,11 +59,6 @@ const TAG_COLORS: &[&str] = &[
     "lightVermilion",
     "lightViolet",
 ];
-
-/// Lists the operations visible to the authenticated user.
-pub fn list_operations(client: &Client) -> Result<Vec<Operation>, HttpError> {
-    client.get_json("/operations")
-}
 
 /// Lists the tags defined for `operation_slug`.
 pub fn list_tags(client: &Client, operation_slug: &str) -> Result<Vec<Tag>, HttpError> {
@@ -133,56 +105,6 @@ mod tests {
 
     fn client_for(server: &MockServer) -> Client {
         Client::new(server.base_url(), test_creds()).expect("client should build")
-    }
-
-    #[test]
-    fn list_operations_deserializes() {
-        let server = MockServer::start();
-        let mock = server.mock(|when, then| {
-            when.method(GET)
-                .path("/api/operations")
-                .header_exists("Authorization")
-                .header_exists("Date");
-            then.status(200)
-                .header("content-type", "application/json")
-                .json_body(serde_json::json!([
-                    { "slug": "s1", "name": "Jack", "numUsers": 1024, "status": 7, "id": 3 },
-                    { "slug": "s2", "name": "Jill", "numUsers": 2048, "status": 2, "id": 10 },
-                ]));
-        });
-
-        let client = client_for(&server);
-        let ops = list_operations(&client).expect("list operations should succeed");
-
-        mock.assert();
-        assert_eq!(ops.len(), 2);
-        assert_eq!(ops[0].slug, "s1");
-        assert_eq!(ops[0].name, "Jack");
-        assert_eq!(ops[0].id, 3);
-        assert_eq!(ops[0].status, 7);
-        assert_eq!(ops[0].num_users, 1024);
-        assert_eq!(ops[1].slug, "s2");
-        assert_eq!(ops[1].num_users, 2048);
-    }
-
-    #[test]
-    fn list_operations_tolerates_missing_optional_fields() {
-        let server = MockServer::start();
-        let mock = server.mock(|when, then| {
-            when.method(GET).path("/api/operations");
-            then.status(200)
-                .header("content-type", "application/json")
-                .json_body(serde_json::json!([{ "slug": "s1", "name": "Jack" }]));
-        });
-
-        let client = client_for(&server);
-        let ops = list_operations(&client).expect("minimal payload should deserialize");
-
-        mock.assert();
-        assert_eq!(ops.len(), 1);
-        assert_eq!(ops[0].slug, "s1");
-        assert_eq!(ops[0].id, 0);
-        assert_eq!(ops[0].num_users, 0);
     }
 
     #[test]
